+++ src/app/werker/werker.page.ts
@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { WerkerService } from 'src/app/werker.service';
 
 @Component({
-  selector: 'app-werker-home',
-  templateUrl: './werker-home.page.html',
-  styleUrls: ['./werker-home.page.scss'],
+  selector: 'app-werker',
+  templateUrl: './werker.page.html',
+  styleUrls: ['./werker.page.scss'],
 })
-export class WerkerHomePage implements OnInit {
+export class WerkerPage implements OnInit {
 
   constructor(
     private werkerService: WerkerService
